Adding overlap support to MeshPartBuilder

+++ lifev/core/array/MapEpetra.hpp
@@ -280,7 +280,7 @@ public:
     //! check if a global id is owned by the current partition
     bool isOwned( const UInt globalId ) const
     {
-        return ( M_uniqueMapEpetra->LID( globalId ) > -1 );
+        return ( M_uniqueMapEpetra->LID( static_cast<int>(globalId) ) > -1 );
     }
 
     //@}
